test(ci): isolate visual accessibility contracts
